requirements.txt
confluent-kafka
pandas
faker
psycopg2-binary
requests
